Printf done

+++ ft_putnbr_base.c
@@ -6,16 +6,16 @@
 /*   By: arpenel <arpenel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 12:03:39 by arpenel           #+#    #+#             */
-/*   Updated: 2025/01/09 14:41:39 by arpenel          ###   ########.fr       */
+/*   Updated: 2025/01/09 15:13:51 by arpenel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_putnbr_base(unsigned int nbr, char *base, size_t *count)
+void	ft_putnbr_base(unsigned long nbr, char *base, size_t *count)
 {
-	long	nb;
-	long	len_base;
+	unsigned long	nb;
+	unsigned long	len_base;
 
 	nb = nbr;
 	len_base = ft_strlen(base);
